create 3 column footer section
