+++ src/app/page.tsx
@@ -1,8 +1,10 @@
 import Footer from "@/components/Footer";
+import GetawaysInspiration from "@/components/GetawaysInspiration";
 
 export default function Home() {
   return (
     <>
+      <GetawaysInspiration />
       <Footer />
     </>
   )
